[Ref] Silence all remaining GCC -Wconversion warnings.
git-svn-id: https://source.openmpt.org/svn/openmpt/trunk/OpenMPT@19289 56274372-70c3-4bfc-bfc3-4c3a0b034d27

+++ soundlib/Load_669.cpp
@@ -225,7 +225,7 @@ bool CSoundFile::Read669(FileReader &file, ModLoadingFlags loadFlags)
 				if(noteInstr <= 0xFE)
 				{
 					m->volcmd = VOLCMD_VOLUME;
-					m->vol = ((vol * 64 + 8) / 15);
+					m->vol = static_cast<uint8>((vol * 64 + 8) / 15);
 				}
 
 				if(effParam != 0xFF)
