+++ soundlib/Load_gdm.cpp
@@ -360,7 +360,7 @@ bool CSoundFile::ReadGDM(FileReader &file, ModLoadingFlags loadFlags)
 					if(note)
 					{
 						note = (note & 0x7F) - 1;  // High bit = no-retrig flag (notes with portamento have this set)
-						m.note = (note & 0x0F) + 12 * (note >> 4) + 12 + NOTE_MIN;
+						m.note = static_cast<uint8>((note & 0x0F) + 12 * (note >> 4) + 12 + NOTE_MIN);
 						if(!m.IsAmigaNote())
 						{
 							onlyAmigaNotes = false;
@@ -483,7 +483,7 @@ bool CSoundFile::ReadGDM(FileReader &file, ModLoadingFlags loadFlags)
 						if(m.command == CMD_S3MCMDEX && ((m.param >> 4) == 0x8) && m.volcmd == VOLCMD_NONE)
 						{
 							m.volcmd = VOLCMD_PANNING;
-							m.vol = ((m.param & 0x0F) * 64 + 8) / 15;
+							m.vol = static_cast<uint8>(((m.param & 0x0F) * 64 + 8) / 15);
 							m.command = oldCmd.command;
 							m.param = oldCmd.param;
 						}
